+++ resources/views/help.blade.php
@@ -1,4 +1,3 @@
-新版本配置结构有变化。<br /><br />The Config structure of this extension was changed in new version.<br /><br />
 <pre>'extensions' => [
             'configx' => [
                 // Set to `false` if you want to disable this extension